get by id,delete mapping

+++ src/main/kotlin/org/openg2p/voucherservice/models/Voucher.kt
@@ -9,7 +9,7 @@ import javax.persistence.*
 data class Voucher(
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    var voucherId          : Long,
+    var voucherId          : Int,
     var voucherCode        : String?,
     @SerializedName("redemptionQuantity")
     var redemptionQuantity : Int,
