+++ src/main/kotlin/org/openg2p/voucherservice/repository/VoucherRepository.kt
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface VoucherRepository : JpaRepository<Voucher, Long>
+interface VoucherRepository : JpaRepository<Voucher, Int>
 {
 }
